requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
openpyxl
xlsxwriter
reportlab
google-genai
python-dotenv
pymysql
sqlalchemy
groq